Finished Inaugrual Project (#1)
* Preliminairy Work

In this commit Question 1 part 1 has been answered. The edgeworth box needs to be optimized.

In the next commit it will be made as a function.

* Updated Jupyter-Notebook (Errors, New Answers, etc)

Minor Errors Found in Notebook:

1. The good have been changed from N = 76, to N = 75.
2. The initial endowments are the reciprocal of each other, so these
are to represent 1 point on the Edgewort box - not 2. Awkward.
3. Added Solution to Question 5 a + b

The Notebook needs to be heavily optimized and cleaned up. It is insanely messy at the current stage.

* # Commit from Pycharm

This commit is a forced commit from pyCharm - without it doesnt allow to switch branches.

Completed Question 6.a. Next commit 6b, 7 and 8.

* # Commit from Pycharm

This commit is a forced commit from pyCharm - without it doesnt allow to switch branches.

Completed Question 6.a. Next commit 6b, 7 and 8.

Question 6b, 7 and 8 is answered.

NOTE: I am not sure how pycharms git extension works.

* # Added Requirements.txt

This file should make it easier to identfy all dependencies.